Changed from double to decimal in a lot of places, made businesses see price without VAT on profile page

+++ VitalHelse/Controllers/SearchController.cs
@@ -80,8 +80,8 @@ public class SearchController : Controller
         List<int> categoryIds,
         int? tagId,
         int? sizeId,
-        double? minPrice,
-        double? maxPrice,
+        decimal? minPrice,
+        decimal? maxPrice,
         bool? discountOnly,
         bool? inStockOnly)
