+++ VitalHelse/Models/Product/TripletexProduct.cs
@@ -7,7 +7,7 @@ public class TripletexProduct
 {
     public int Id { get; set; }
     public string Name { get; set; } = "";
-    public double PriceExVat { get; set; }
-    public double PriceInVat { get; set; }
+    public decimal PriceExVat { get; set; }
+    public decimal PriceInVat { get; set; }
     public double StockCount { get; set; }
 }
